adapter: Move main.go to '/cmd'
